requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-appconf==1.0.2
django-blog-zinnia==0.19
django-ckeditor==5.4.0
django-constance==2.1.0
django-contrib-comments==1.8.0
django-debug-toolbar==1.9.1
django-imagekit==4.0.2
django-jet==1.0.7
django-js-asset==1.0.0
django-material==1.2.2
django-modeladmin-reorder==0.3
django-modeltranslation==0.12.1
django-mptt==0.9.0
django-picklefield==1.0.0
django-rosetta==0.8.1
django-tagging==0.4.6
django-xmlrpc==0.1.8
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.10.3
idna==2.6
microsofttranslator==0.8
mock==2.0.0
mots-vides==2015.5.11
oauth2client==4.1.2
pbr==3.1.1
pilkit==2.0
Pillow==5.0.0
polib==1.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyparsing==2.2.0
python-memcached==1.59
pytz==2018.3
regex==2018.2.21
requests==2.18.4
rsa==3.4.2
six==1.11.0
sqlparse==0.2.4
Unipath==1.1
uritemplate==3.0.0
urllib3==1.22
